вынесены две функции  в утилитарные для переключения селекторов классов
